fix(search): accept a fractional runtime on the search request
The unmatched scan row carries ffprobe's container duration, which is
fractional. The search request bound it to int?, so System.Text.Json threw
on every library-import match attempt and the whole body was rejected with
a 400. Nothing matched, and the log only said "Error parsing search request
body".

DurationSeconds is now double? on the wire, matching the producer and the
frontend type that was already number. The scorer still works in whole
minutes, so the workflow rounds on the way in and drops a runtime it cannot
use rather than feeding the ranking a bogus one.

+++ listenarr.application/Search/Core/SearchRequest.cs
@@ -59,7 +59,9 @@ namespace Listenarr.Application.Search.Core
 
         // Known runtime of the files being matched, in seconds. Used to rank candidates: it is
         // what separates Mistborn (1499 min) from Mistborn: Secret History (329 min).
-        public int? DurationSeconds { get; set; }
+        // Fractional, because it comes from ffprobe's container duration via the unmatched scan
+        // row, which reports seconds to two decimal places.
+        public double? DurationSeconds { get; set; }
     }
 
     // Options to control MyAnonamouse (MyAnonamouse.net) search behavior
